requirements.txt
termcolor
requests
BeautifulSoup4
websocket-client
html5lib
lxml
